Add wavelet-domain image CS experiment results for Tree-HyperLISTA
TH-hybrid (3 params) achieves 31.66 dB PSNR at CS 50%, beating HyperLISTA
(21.56 dB) by 10 dB and outperforming all other 3-param methods. Fixed
training data generation to align tree structure with wavelet coefficient
hierarchy. Added block baselines (SH-topk, Group-FISTA) for comparison.

+++ experiments/run_tree_image_cs.py
@@ -68,8 +68,8 @@ def generate_wavelet_tree_sparse_data(tree_info, A, num_samples, target_K,
     Y = Y_clean + noise
 
     return {
-        'x': torch.from_numpy(X),
-        'y': torch.from_numpy(Y),
+        'x': torch.from_numpy(X).float(),
+        'y': torch.from_numpy(Y.astype(np.float32)).float(),
         'noise_std': sigma,
     }
